Refactor image storage

+++ src/main/java/com/aemreunal/helper/ImageStorage.java
@@ -16,33 +16,27 @@ package com.aemreunal.helper;
  ***************************
  */
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.util.UUID;
 import org.springframework.web.multipart.MultipartFile;
 import com.aemreunal.config.GlobalSettings;
 import com.aemreunal.exception.region.MapImageDeleteException;
+import com.aemreunal.exception.region.MapImageLoadException;
 import com.aemreunal.exception.region.MapImageSaveException;
 import com.aemreunal.exception.region.MultipartFileReadException;
 
 public class ImageStorage {
 
-    public String saveImage(String username, Long projectId, Long regionId, String existingImageName, MultipartFile imageFile) throws MultipartFileReadException, MapImageDeleteException, MapImageSaveException {
-        byte[] fileBytes;
-        try {
-            fileBytes = imageFile.getBytes();
-        } catch (IOException e) {
-            throw new MultipartFileReadException(projectId, regionId);
-        }
-        return saveImage(username, projectId, regionId, existingImageName, fileBytes);
-    }
-
     /**
-     * Saves the given image (in bytes) to the filesystem and returns the name of the
-     * saved image file.
+     * Saves the given image to the filesystem and returns the name of the saved image
+     * file.
      * <p>
      * The method saves images under the home folder of the user, inside the <code>{@value
-     * com.aemreunal.config.GlobalSettings#ROOT_STORAGE_FOLDER_DIRECTORY_NAME}</code> root
+     * com.aemreunal.config.GlobalSettings#ROOT_STORAGE_FOLDER_DIRECTORY_NAME}</code>
      * folder, <code>{@value com.aemreunal.config.GlobalSettings#IMAGE_STORAGE_FOLDER_DIRECTORY_NAME}</code>
      * sub folder. If the folder does not exist, it will be created.
      *
@@ -57,24 +51,19 @@ public class ImageStorage {
      *         If an image of a similar function already exists, the name of that image
      *         file should be given with this parameter. When this parameter is not
      *         <code>null</code>, this file is deleted prior to saving the new one.
-     * @param imageAsBytes
-     *         The image as a <code>byte</code> array.
+     * @param imageMultipartFile
+     *         The image as a {@link org.springframework.web.multipart.MultipartFile
+     *         MultipartFile}.
      *
      * @return The name of the saved image file.
      */
-    public String saveImage(String username, Long projectId, Long regionId, String existingImageName, byte[] imageAsBytes)
-            throws MapImageSaveException, MapImageDeleteException {
+    public String saveImage(String username, Long projectId, Long regionId, String existingImageName, MultipartFile imageMultipartFile) throws MultipartFileReadException, MapImageDeleteException, MapImageSaveException {
         File imageFile;
         String fileName;
         String filePath = getFilePath(username, projectId, regionId);
 
         // Delete pre-existing image file
-        if (existingImageName != null) {
-            if (!deleteImage(username, projectId, regionId, existingImageName)) {
-                System.err.println("Unable to delete existing image!");
-                throw new MapImageDeleteException(projectId, regionId);
-            }
-        }
+        deleteImage(username, projectId, regionId, existingImageName);
 
         // Ensure unique file name
         do {
@@ -84,62 +73,46 @@ public class ImageStorage {
 
         // Check for the existence of the parent folder and create
         // if it doesn't exist
-        if (!createParentFolder(imageFile)) {
-            System.err.println("Unable to create parent folder!");
-            throw new MapImageSaveException(projectId, regionId);
-        }
+        createParentFolder(projectId, regionId, imageFile);
 
         // Create the new image file
-        if (!createFile(imageFile)) {
-            System.err.println("Unable to create file!");
-            throw new MapImageSaveException(projectId, regionId);
-        }
+        createFile(projectId, regionId, imageFile);
 
         // Write the image bytes to the new file
-        if (!writeImageToFile(imageAsBytes, imageFile)) {
-            System.err.println("Unable to write image to file!");
-            throw new MapImageSaveException(projectId, regionId);
-        }
+        writeImageToFile(projectId, regionId, imageMultipartFile, imageFile);
+
         return fileName;
     }
 
-    private boolean createParentFolder(File imageFile) {
+    private void createParentFolder(Long projectId, Long regionId, File imageFile) throws MapImageSaveException {
         if (!imageFile.getParentFile().exists()) {
             // If it doesn't exist, create it
             if (!imageFile.getParentFile().mkdirs()) {
-                return false;
+                System.err.println("Unable to create parent folders!");
+                throw new MapImageSaveException(projectId, regionId);
             }
         }
-        return true;
     }
 
-    private boolean createFile(File imageFile) {
+    private void createFile(Long projectId, Long regionId, File imageFile) throws MapImageSaveException {
         try {
             if (!imageFile.createNewFile()) {
                 System.err.println("Unable to createNewFile()!");
-                return false;
+                throw new MapImageSaveException(projectId, regionId);
             }
         } catch (IOException e) {
             System.err.println("Unable to create file!");
-            e.printStackTrace();
-            return false;
+            throw new MapImageSaveException(projectId, regionId);
         }
-        return true;
     }
 
-    private boolean writeImageToFile(byte[] imageAsBytes, File imageFile) {
+    private void writeImageToFile(Long projectId, Long regionId, MultipartFile imageMultipartFile, File imageFile) throws MapImageSaveException {
         try {
-            FileOutputStream stream = new FileOutputStream(imageFile);
-            stream.write(imageAsBytes);
-            stream.close();
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-            return false;
+            imageMultipartFile.transferTo(imageFile);
         } catch (IOException e) {
-            e.printStackTrace();
-            return false;
+            System.err.println("Unable to write image to file!");
+            throw new MapImageSaveException(projectId, regionId);
         }
-        return true;
     }
 
     /**
@@ -159,40 +132,37 @@ public class ImageStorage {
      * @return A <code>byte[]</code> if the image file has been successfully loaded and
      * read, <code>null</code> otherwise.
      */
-    public byte[] loadImage(String username, Long projectId, Long regionId, String imageFileName) {
+    public byte[] loadImage(String username, Long projectId, Long regionId, String imageFileName) throws MapImageLoadException {
         String filePath = getFilePath(username, projectId, regionId);
         File imageFile = new File(filePath + imageFileName);
         if (!imageFile.exists()) {
             System.err.println("Image file does not exist!");
-            return null;
+            throw new MapImageLoadException(projectId, regionId);
         }
+        return loadImageFromFile(projectId, regionId, imageFile);
+    }
+
+    private byte[] loadImageFromFile(Long projectId, Long regionId, File imageFile) throws MapImageLoadException {
         // Can safely cast from long to int, as image file will never exceed
         // 2^32 bytes (Which would've used 64 bits of long).
         byte[] imageAsBytes = new byte[(int) imageFile.length()];
-        if (!loadImageFromFile(imageFile, imageAsBytes)) {
-            System.err.println("Unable to load image file!");
-            return null;
-        }
-        return imageAsBytes;
-    }
-
-    private boolean loadImageFromFile(File imageFile, byte[] imageAsBytes) {
         try {
             FileInputStream stream = new FileInputStream(imageFile);
             stream.read(imageAsBytes);
             stream.close();
         } catch (FileNotFoundException e) {
-            e.printStackTrace();
-            return false;
+            System.err.println("File to read from is not found!");
+            throw new MapImageLoadException(projectId, regionId);
         } catch (IOException e) {
-            e.printStackTrace();
-            return false;
+            System.err.println("Unable to read from file!");
+            throw new MapImageLoadException(projectId, regionId);
         }
-        return true;
+        return imageAsBytes;
     }
 
     /**
-     * Deletes the specified image file.
+     * Deletes the specified image file. The method will do nothing if a {@code null}
+     * value is provided for {@code imageFileName} parameter.
      *
      * @param username
      *         The username of whoever is deleting the image.
@@ -202,21 +172,21 @@ public class ImageStorage {
      * @param regionId
      *         The ID of the region of the image.
      * @param imageFileName
-     *         The name of the image file to be deleted.
-     *
-     * @return <code>true</code> if image file has been successfully deleted,
-     * <code>false</code> otherwise.
+     *         The name of the image file to be deleted. If a {@code null} value is
+     *         provided, the method will do nothing and return.
      */
-    public boolean deleteImage(String username, Long projectId, Long regionId, String imageFileName) {
+    public void deleteImage(String username, Long projectId, Long regionId, String imageFileName) throws MapImageDeleteException {
+        if (imageFileName == null) {
+            return;
+        }
         String filePath = getFilePath(username, projectId, regionId);
         File imageFile = new File(filePath + imageFileName);
         try {
             Files.delete(imageFile.toPath());
         } catch (IOException e) {
-            e.printStackTrace();
-            return false;
+            System.err.println("Unable to delete the image!");
+            throw new MapImageDeleteException(projectId, regionId);
         }
-        return true;
     }
 
     private String getFilePath(String username, Long projectId, Long regionId) {
